+++ Testing/TinyFFR.Tests/Assets/Meshes/MeshConfig.cs
@@ -115,6 +115,7 @@ class MeshConfigTest {
 			LinearRescalingFactor = 123f,
 			BoundingBoxOverride = null,
 			BoundingBoxAdditionalMargin = 0f,
+			AllowsPerInstanceVertexMutation = true,
 			Name = "Aa Aa"
 		};
 		var testConfigB = new MeshCreationConfig {
@@ -125,6 +126,7 @@ class MeshConfigTest {
 			LinearRescalingFactor = -0.123f,
 			BoundingBoxOverride = PositionedCuboid.FromHalfDimensions(1f, 2f, 3f, new Location(4f, 5f, 6f)),
 			BoundingBoxAdditionalMargin = 0.5f,
+			AllowsPerInstanceVertexMutation = false,
 			Name = "BBBbbb"
 		};
 
@@ -136,6 +138,7 @@ class MeshConfigTest {
 			Assert.AreEqual(expected.LinearRescalingFactor, actual.LinearRescalingFactor);
 			Assert.AreEqual(expected.BoundingBoxOverride, actual.BoundingBoxOverride);
 			Assert.AreEqual(expected.BoundingBoxAdditionalMargin, actual.BoundingBoxAdditionalMargin);
+			Assert.AreEqual(expected.AllowsPerInstanceVertexMutation, actual.AllowsPerInstanceVertexMutation);
 			Assert.AreEqual(expected.Name.ToString(), actual.Name.ToString());
 		}
 
@@ -151,6 +154,7 @@ class MeshConfigTest {
 			.Bool(false)
 			.Obj(default(PositionedCuboid))
 			.Float(0f)
+			.Bool(true)
 			.String("Aa Aa")
 			.For(testConfigA);
 
@@ -163,6 +167,7 @@ class MeshConfigTest {
 			.Bool(true)
 			.Obj(PositionedCuboid.FromHalfDimensions(1f, 2f, 3f, new Location(4f, 5f, 6f)))
 			.Float(0.5f)
+			.Bool(false)
 			.String("BBBbbb")
 			.For(testConfigB);
 
@@ -174,6 +179,7 @@ class MeshConfigTest {
 			.Including(nameof(MeshCreationConfig.LinearRescalingFactor))
 			.Including(nameof(MeshCreationConfig.BoundingBoxOverride))
 			.Including(nameof(MeshCreationConfig.BoundingBoxAdditionalMargin))
+			.Including(nameof(MeshCreationConfig.AllowsPerInstanceVertexMutation))
 			.Including(nameof(MeshCreationConfig.Name))
 			.End();
 	}