+++ TinyFFR.Native/tffr/headers/scene/objects/native_impl_objects.h
@@ -11,7 +11,8 @@ typedef uintptr_t ModelInstanceHandle;
 class native_impl_objects {
 public:
 	static void allocate_model_instance(mat4f* initialTransformPtr, VertexBufferHandle vb, IndexBufferHandle ib, int32_t ibStartIndex, int32_t ibCount, int32_t boneCount, MaterialHandle material, float3 aabbCenter, float3 aabbHalfExtents, ModelInstanceHandle* outModelInstance);
-	static void set_model_instance_mesh(ModelInstanceHandle modelInstance, VertexBufferHandle vb, IndexBufferHandle ib, int32_t ibStartIndex, int32_t ibCount, float3 aabbCenter, float3 aabbHalfExtents);
+	static void set_model_instance_mesh(ModelInstanceHandle modelInstance, VertexBufferHandle vb, IndexBufferHandle ib, int32_t ibStartIndex, int32_t ibCount);
+	static void set_model_instance_aabb(ModelInstanceHandle modelInstance, float3 aabbCenter, float3 aabbHalfExtents);
 	static void set_model_instance_material(ModelInstanceHandle modelInstance, MaterialHandle material);
 	static void set_model_instance_world_mat(ModelInstanceHandle modelInstance, mat4f* worldMatPtr);
 	static void set_model_instance_bone_transforms(ModelInstanceHandle modelInstance, mat4f* transforms, int32_t boneCount);